elixirDoc v0.0.1 ArrayUtils

Array manipulation utils

Do you need to calculate ?

If you need to operate on numbers, refer to Calculator. For instance you can use Calculator.add/2. You also have access to this type Calculator.number_and_string/0

Functions

removeHead(list)

Returns an array with all elements buts the first

Examples

iex> ArrayUtils.removeHead([1,2])

[2]